Migrate useMediaQuery to ts
