+++ lib/hooks/index.js
@@ -1,5 +1,4 @@
 export { useSortEpisodes } from './useSortEpisodes'
 export { useMediaControls } from './useMediaControls'
 export { useMediaSessions } from './useMediaSessions'
-export { useMediaQuery, useIsMobile } from './useMediaQueries'
 export { useFetch } from './useFetch'
