+++ pages/index.tsx
@@ -2,7 +2,7 @@ import { InferGetStaticPropsType } from 'next'
 import dynamic from 'next/dynamic'
 import { getRecommendedChannels } from 'lib/api'
 import { useFavs } from 'lib/hooks/use-favs'
-import { useIsMobile } from 'lib/hooks'
+import { useIsMobile } from 'lib/hooks/use-media-queries'
 import Layout from 'components/common/Layout'
 import MainTitle from 'components/MainTitle'
 import { ChannelsGrid } from 'components/channel/ChannelsGrid'
